src/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepseek>=1.0.0",
    "load-dotenv>=0.1.0",
    "openai>=1.61.0",
    "python-telegram-bot>=21.10",
    "requests>=2.32.3",
    "web3>=7.8.0",
]
